# Adds bcrypt native binding for better login performance