--- a/package.json
+++ b/package.json
@@ -77,5 +77,8 @@
   },
   "bin": {
     "parse-server": "./bin/parse-server"
+  },
+  "optionalDependencies": {
+    "bcrypt": "0.8.7"
   }
 }