--- a/src/password.js
+++ b/src/password.js
@@ -1,9 +1,8 @@
 // Tools for encrypting and decrypting passwords.
 // Basically promise-friendly wrappers for bcrypt.
 var bcrypt = require('bcrypt-nodejs');
-
 // Returns a promise for a hashed password string.
-function hash(password) {
+var hash = function(password) {
   return new Promise(function(fulfill, reject) {
     bcrypt.hash(password, null, null, function(err, hashedPassword) {
       if (err) {
@@ -14,6 +13,21 @@ function hash(password) {
     });
   });
 }
+
+try {
+  bcrypt = require('bcrypt');
+  hash = function(password) {
+    return new Promise(function(fulfill, reject) {
+      bcrypt.hash(password, 10,function(err, hashedPassword) {
+        if (err) {
+          reject(err);
+        } else {
+          fulfill(hashedPassword);
+        }
+      });
+    });
+  }
+} catch(e) {}
 
 // Returns a promise for whether this password compares to equal this
 // hashed password.