Fixes #2539 

- Adds optional bcrypt native binding for better performance
- Defaults salt rounds to 10 


> Well... Here are some benchmarks. Ran with `ab -n200 -c100` locally 5 times for each module.

* **Baseline** (immediately fulfill promises in password.js without bcrypt use): 250 req/sec
* **[bcrypt-nodejs](https://github.com/shaneGirish/bcrypt-nodejs)** (current implementation): 5 req/sec
* **[bcryptjs](https://github.com/dcodeIO/bcrypt.js)**: 5 req/sec
* **[twin-bcrypt](https://github.com/fpirsch/twin-bcrypt)**: 5 req/sec
* **[bcrypt](https://github.com/ncb000gt/node.bcrypt.js)** (native bindings): 52 req/sec