@JeremyPlease 2016-08-19
@flovilmart It's important that we use bcrypt native bindings for the compare function too.